+++ crates/client/db/src/lib.rs
@@ -25,6 +25,7 @@ pub mod class_db;
 pub mod contract_db;
 pub mod db_block_id;
 pub mod db_metrics;
+pub mod l1_db;
 pub mod storage_updates;
 
 pub use error::{DeoxysStorageError, TrieType};
@@ -164,6 +165,9 @@ pub enum Column {
     BonsaiClassesTrie,
     BonsaiClassesFlat,
     BonsaiClassesLog,
+
+    L1Messaging,
+    L1MessagingNonce,
 }
 
 impl fmt::Debug for Column {
@@ -207,6 +211,8 @@ impl Column {
             BonsaiClassesTrie,
             BonsaiClassesFlat,
             BonsaiClassesLog,
+            L1Messaging,
+            L1MessagingNonce,
             PendingContractToClassHashes,
             PendingContractToNonces,
             PendingContractStorage,
@@ -242,6 +248,8 @@ impl Column {
             ContractToNonces => "contract_to_nonces",
             ContractClassHashes => "contract_class_hashes",
             ContractStorage => "contract_storage",
+            L1Messaging => "l1_messaging",
+            L1MessagingNonce => "l1_messaging_nonce",
             PendingContractToClassHashes => "pending_contract_to_class_hashes",
             PendingContractToNonces => "pending_contract_to_nonces",
             PendingContractStorage => "pending_contract_storage",