+++ crates/primitives/transactions/src/lib.rs
@@ -2,21 +2,18 @@ mod broadcasted_to_blockifier;
 pub mod compute_hash;
 mod from_blockifier;
 mod from_broadcasted_transaction;
-mod from_starknet_api;
 mod from_starknet_provider;
-mod to_starknet_api;
+mod into_starknet_api;
 mod to_starknet_core;
 pub mod utils;
 
-use blockifier::transaction::objects::FeeType;
 pub use broadcasted_to_blockifier::broadcasted_to_blockifier;
 use dp_convert::ToFelt;
 pub use from_starknet_provider::TransactionTypeError;
 use starknet_api::transaction::TransactionVersion;
 use starknet_types_core::{felt::Felt, hash::StarkHash};
 
-const SIMULATE_TX_VERSION_OFFSET: Felt =
-    Felt::from_raw([576460752142434320, 18446744073709551584, 17407, 18446744073700081665]);
+const SIMULATE_TX_VERSION_OFFSET: Felt = Felt::from_hex_unchecked("0x100000000000000000000000000000000");
 
 /// Legacy check for deprecated txs
 /// See `https://docs.starknet.io/documentation/architecture_and_concepts/Blocks/transactions/` for more details.
@@ -115,6 +112,65 @@ impl From<DeployAccountTransactionV3> for Transaction {
     }
 }
 
+#[derive(Debug, thiserror::Error)]
+#[error("Transaction type is note the expected one.")]
+pub struct UnexpectedTransactionType;
+
+impl TryFrom<Transaction> for InvokeTransaction {
+    type Error = UnexpectedTransactionType;
+
+    fn try_from(tx: Transaction) -> Result<Self, Self::Error> {
+        match tx {
+            Transaction::Invoke(tx) => Ok(tx),
+            _ => Err(UnexpectedTransactionType),
+        }
+    }
+}
+
+impl TryFrom<Transaction> for L1HandlerTransaction {
+    type Error = UnexpectedTransactionType;
+
+    fn try_from(tx: Transaction) -> Result<Self, Self::Error> {
+        match tx {
+            Transaction::L1Handler(tx) => Ok(tx),
+            _ => Err(UnexpectedTransactionType),
+        }
+    }
+}
+
+impl TryFrom<Transaction> for DeclareTransaction {
+    type Error = UnexpectedTransactionType;
+
+    fn try_from(tx: Transaction) -> Result<Self, Self::Error> {
+        match tx {
+            Transaction::Declare(tx) => Ok(tx),
+            _ => Err(UnexpectedTransactionType),
+        }
+    }
+}
+
+impl TryFrom<Transaction> for DeployTransaction {
+    type Error = UnexpectedTransactionType;
+
+    fn try_from(tx: Transaction) -> Result<Self, Self::Error> {
+        match tx {
+            Transaction::Deploy(tx) => Ok(tx),
+            _ => Err(UnexpectedTransactionType),
+        }
+    }
+}
+
+impl TryFrom<Transaction> for DeployAccountTransaction {
+    type Error = UnexpectedTransactionType;
+
+    fn try_from(tx: Transaction) -> Result<Self, Self::Error> {
+        match tx {
+            Transaction::DeployAccount(tx) => Ok(tx),
+            _ => Err(UnexpectedTransactionType),
+        }
+    }
+}
+
 impl Transaction {
     pub fn version(&self) -> TransactionVersion {
         match self {
@@ -135,11 +191,11 @@ impl Transaction {
         !matches!(self, Transaction::L1Handler(_))
     }
 
-    pub fn fee_type(&self) -> FeeType {
+    pub fn fee_type(&self) -> blockifier::transaction::objects::FeeType {
         if self.is_l1_handler() || self.version() < TransactionVersion::THREE {
-            FeeType::Eth
+            blockifier::transaction::objects::FeeType::Eth
         } else {
-            FeeType::Strk
+            blockifier::transaction::objects::FeeType::Strk
         }
     }
 }
@@ -149,6 +205,24 @@ pub enum InvokeTransaction {
     V0(InvokeTransactionV0),
     V1(InvokeTransactionV1),
     V3(InvokeTransactionV3),
+}
+
+impl From<InvokeTransactionV0> for InvokeTransaction {
+    fn from(tx: InvokeTransactionV0) -> Self {
+        InvokeTransaction::V0(tx)
+    }
+}
+
+impl From<InvokeTransactionV1> for InvokeTransaction {
+    fn from(tx: InvokeTransactionV1) -> Self {
+        InvokeTransaction::V1(tx)
+    }
+}
+
+impl From<InvokeTransactionV3> for InvokeTransaction {
+    fn from(tx: InvokeTransactionV3) -> Self {
+        InvokeTransaction::V3(tx)
+    }
 }
 
 impl InvokeTransaction {
@@ -179,11 +253,11 @@ impl InvokeTransaction {
         H::hash_array(self.signature())
     }
 
-    pub fn calldata(&self) -> Option<&[Felt]> {
+    pub fn calldata(&self) -> &[Felt] {
         match self {
-            InvokeTransaction::V0(tx) => Some(&tx.calldata),
-            InvokeTransaction::V1(tx) => Some(&tx.calldata),
-            InvokeTransaction::V3(tx) => Some(&tx.calldata),
+            InvokeTransaction::V0(tx) => &tx.calldata,
+            InvokeTransaction::V1(tx) => &tx.calldata,
+            InvokeTransaction::V3(tx) => &tx.calldata,
         }
     }
 
@@ -281,6 +355,30 @@ pub enum DeclareTransaction {
     V3(DeclareTransactionV3),
 }
 
+impl From<DeclareTransactionV0> for DeclareTransaction {
+    fn from(tx: DeclareTransactionV0) -> Self {
+        DeclareTransaction::V0(tx)
+    }
+}
+
+impl From<DeclareTransactionV1> for DeclareTransaction {
+    fn from(tx: DeclareTransactionV1) -> Self {
+        DeclareTransaction::V1(tx)
+    }
+}
+
+impl From<DeclareTransactionV2> for DeclareTransaction {
+    fn from(tx: DeclareTransactionV2) -> Self {
+        DeclareTransaction::V2(tx)
+    }
+}
+
+impl From<DeclareTransactionV3> for DeclareTransaction {
+    fn from(tx: DeclareTransactionV3) -> Self {
+        DeclareTransaction::V3(tx)
+    }
+}
+
 impl DeclareTransaction {
     fn version(&self) -> TransactionVersion {
         match self {
@@ -313,10 +411,6 @@ impl DeclareTransaction {
         H: StarkHash,
     {
         H::hash_array(self.signature())
-    }
-
-    pub fn call_data(&self) -> Option<&[Felt]> {
-        None
     }
 
     pub fn nonce(&self) -> &Felt {
@@ -415,6 +509,18 @@ pub enum DeployAccountTransaction {
     V3(DeployAccountTransactionV3),
 }
 
+impl From<DeployAccountTransactionV1> for DeployAccountTransaction {
+    fn from(tx: DeployAccountTransactionV1) -> Self {
+        DeployAccountTransaction::V1(tx)
+    }
+}
+
+impl From<DeployAccountTransactionV3> for DeployAccountTransaction {
+    fn from(tx: DeployAccountTransactionV3) -> Self {
+        DeployAccountTransaction::V3(tx)
+    }
+}
+
 impl DeployAccountTransaction {
     pub fn version(&self) -> TransactionVersion {
         match self {
@@ -443,10 +549,10 @@ impl DeployAccountTransaction {
         H::hash_array(self.signature())
     }
 
-    pub fn calldata(&self) -> Option<&[Felt]> {
+    pub fn calldata(&self) -> &[Felt] {
         match self {
-            DeployAccountTransaction::V1(tx) => Some(&tx.constructor_calldata),
-            DeployAccountTransaction::V3(tx) => Some(&tx.constructor_calldata),
+            DeployAccountTransaction::V1(tx) => &tx.constructor_calldata,
+            DeployAccountTransaction::V3(tx) => &tx.constructor_calldata,
         }
     }
 
@@ -558,6 +664,415 @@ impl From<starknet_core::types::DataAvailabilityMode> for DataAvailabilityMode {
         match da_mode {
             starknet_core::types::DataAvailabilityMode::L1 => Self::L1,
             starknet_core::types::DataAvailabilityMode::L2 => Self::L2,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_tx_with_hash() {
+        let tx: Transaction = InvokeTransactionV0::default().into();
+        let hash = Felt::from_hex_unchecked("0x1234567890abcdef");
+        let tx_with_hash = TransactionWithHash::new(tx.clone(), hash);
+        assert_eq!(tx_with_hash.transaction, tx);
+        assert_eq!(tx_with_hash.hash, hash);
+    }
+
+    #[test]
+    fn test_tx_version() {
+        let tx: Transaction = InvokeTransactionV0::default().into();
+        assert_eq!(tx.version(), TransactionVersion::ZERO);
+
+        let tx: Transaction = InvokeTransactionV1::default().into();
+        assert_eq!(tx.version(), TransactionVersion::ONE);
+
+        let tx: Transaction = InvokeTransactionV3::default().into();
+        assert_eq!(tx.version(), TransactionVersion::THREE);
+
+        let tx: Transaction = L1HandlerTransaction::default().into();
+        assert_eq!(tx.version(), TransactionVersion::ZERO);
+
+        let tx: Transaction = DeclareTransactionV0::default().into();
+        assert_eq!(tx.version(), TransactionVersion::ZERO);
+
+        let tx: Transaction = DeclareTransactionV1::default().into();
+        assert_eq!(tx.version(), TransactionVersion::ONE);
+
+        let tx: Transaction = DeclareTransactionV2::default().into();
+        assert_eq!(tx.version(), TransactionVersion::TWO);
+
+        let tx: Transaction = DeclareTransactionV3::default().into();
+        assert_eq!(tx.version(), TransactionVersion::THREE);
+
+        let tx: Transaction = DeployTransaction::default().into();
+        assert_eq!(tx.version(), TransactionVersion::ZERO);
+
+        let tx: Transaction = DeployAccountTransactionV1::default().into();
+        assert_eq!(tx.version(), TransactionVersion::ONE);
+
+        let tx: Transaction = DeployAccountTransactionV3::default().into();
+        assert_eq!(tx.version(), TransactionVersion::THREE);
+    }
+
+    #[test]
+    fn test_try_from_tx() {
+        let invoke_tx: InvokeTransaction = dummy_tx_invoke_v0().into();
+        let tx = Transaction::Invoke(invoke_tx.clone());
+        assert_eq!(InvokeTransaction::try_from(tx.clone()).unwrap(), invoke_tx);
+
+        let l1_handler_tx: L1HandlerTransaction = dummy_l1_handler();
+        let tx = Transaction::L1Handler(l1_handler_tx.clone());
+        assert_eq!(L1HandlerTransaction::try_from(tx.clone()).unwrap(), l1_handler_tx);
+
+        let declare_tx: DeclareTransaction = dummy_tx_declare_v0().into();
+        let tx = Transaction::Declare(declare_tx.clone());
+        assert_eq!(DeclareTransaction::try_from(tx.clone()).unwrap(), declare_tx);
+
+        let deploy_tx: DeployTransaction = dummy_tx_deploy();
+        let tx = Transaction::Deploy(deploy_tx.clone());
+        assert_eq!(DeployTransaction::try_from(tx.clone()).unwrap(), deploy_tx);
+
+        let deploy_account_tx: DeployAccountTransaction = dummy_tx_deploy_account_v1().into();
+        let tx = Transaction::DeployAccount(deploy_account_tx.clone());
+        assert_eq!(DeployAccountTransaction::try_from(tx.clone()).unwrap(), deploy_account_tx);
+    }
+
+    #[test]
+    fn test_tx_is_l1_handler() {
+        let tx: Transaction = L1HandlerTransaction::default().into();
+        assert!(tx.is_l1_handler());
+        let tx: Transaction = InvokeTransactionV0::default().into();
+        assert!(!tx.is_l1_handler());
+    }
+
+    #[test]
+    fn test_tx_is_account() {
+        let tx: Transaction = L1HandlerTransaction::default().into();
+        assert!(!tx.is_account());
+
+        let tx: Transaction = InvokeTransactionV0::default().into();
+        assert!(tx.is_account());
+    }
+
+    #[test]
+    fn test_tx_fee_type() {
+        let tx: Transaction = L1HandlerTransaction::default().into();
+        assert!(matches!(tx.fee_type(), blockifier::transaction::objects::FeeType::Eth));
+
+        let tx: Transaction = InvokeTransactionV0::default().into();
+        assert!(matches!(tx.fee_type(), blockifier::transaction::objects::FeeType::Eth));
+
+        let tx: Transaction = InvokeTransactionV3::default().into();
+        assert!(matches!(tx.fee_type(), blockifier::transaction::objects::FeeType::Strk));
+    }
+
+    #[test]
+    fn test_sender_address() {
+        let tx: InvokeTransaction = dummy_tx_invoke_v0().into();
+        assert_eq!(tx.sender_address(), &Felt::from(4));
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v1().into();
+        assert_eq!(tx.sender_address(), &Felt::from(1));
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v3().into();
+        assert_eq!(tx.sender_address(), &Felt::from(1));
+
+        let tx: DeclareTransaction = dummy_tx_declare_v0().into();
+        assert_eq!(tx.sender_address(), &Felt::from(1));
+
+        let tx: DeclareTransaction = dummy_tx_declare_v1().into();
+        assert_eq!(tx.sender_address(), &Felt::from(1));
+
+        let tx: DeclareTransaction = dummy_tx_declare_v2().into();
+        assert_eq!(tx.sender_address(), &Felt::from(1));
+
+        let tx: DeclareTransaction = dummy_tx_declare_v3().into();
+        assert_eq!(tx.sender_address(), &Felt::from(1));
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v1().into();
+        assert_eq!(tx.sender_address(), &Felt::from(5));
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v3().into();
+        assert_eq!(tx.sender_address(), &Felt::from(4));
+    }
+
+    #[test]
+    fn test_signature() {
+        let tx: InvokeTransaction = dummy_tx_invoke_v0().into();
+        assert_eq!(tx.signature(), &[Felt::from(2), Felt::from(3)]);
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v1().into();
+        assert_eq!(tx.signature(), &[Felt::from(5), Felt::from(6)]);
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v3().into();
+        assert_eq!(tx.signature(), &[Felt::from(4), Felt::from(5)]);
+
+        let tx: DeclareTransaction = dummy_tx_declare_v0().into();
+        assert_eq!(tx.signature(), &[Felt::from(3), Felt::from(4)]);
+
+        let tx: DeclareTransaction = dummy_tx_declare_v1().into();
+        assert_eq!(tx.signature(), &[Felt::from(3), Felt::from(4)]);
+
+        let tx: DeclareTransaction = dummy_tx_declare_v2().into();
+        assert_eq!(tx.signature(), &[Felt::from(4), Felt::from(5)]);
+
+        let tx: DeclareTransaction = dummy_tx_declare_v3().into();
+        assert_eq!(tx.signature(), &[Felt::from(3), Felt::from(4)]);
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v1().into();
+        assert_eq!(tx.signature(), &[Felt::from(2), Felt::from(3)]);
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v3().into();
+        assert_eq!(tx.signature(), &[Felt::from(1), Felt::from(2)]);
+    }
+
+    #[test]
+    fn test_compute_hash_signature() {
+        let tx: InvokeTransaction = dummy_tx_invoke_v0().into();
+        assert_eq!(
+            tx.compute_hash_signature::<starknet_types_core::hash::Poseidon>(),
+            Felt::from_hex_unchecked("0xcb40676fdafc07998f1391bf66356b5f0c96d5e61bbf5e7ec0aa0320dd8f9b")
+        );
+
+        let tx: DeclareTransaction = dummy_tx_declare_v0().into();
+        assert_eq!(
+            tx.compute_hash_signature::<starknet_types_core::hash::Poseidon>(),
+            Felt::from_hex_unchecked("0x22d481b177090ea8db58ceece7d8493e746d690a1708d438c6c4e51b23c81ee")
+        );
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v1().into();
+        assert_eq!(
+            tx.compute_hash_signature::<starknet_types_core::hash::Poseidon>(),
+            Felt::from_hex_unchecked("0xcb40676fdafc07998f1391bf66356b5f0c96d5e61bbf5e7ec0aa0320dd8f9b")
+        );
+    }
+
+    #[test]
+    fn test_calldata() {
+        let tx: InvokeTransaction = dummy_tx_invoke_v0().into();
+        assert_eq!(tx.calldata(), &[Felt::from(6), Felt::from(7)]);
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v1().into();
+        assert_eq!(tx.calldata(), &[Felt::from(2), Felt::from(3)]);
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v3().into();
+        assert_eq!(tx.calldata(), &[Felt::from(2), Felt::from(3)]);
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v1().into();
+        assert_eq!(tx.calldata(), &[Felt::from(6), Felt::from(7)]);
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v3().into();
+        assert_eq!(tx.calldata(), &[Felt::from(5), Felt::from(6)]);
+    }
+
+    #[test]
+    pub fn test_nonce() {
+        let tx: InvokeTransaction = dummy_tx_invoke_v0().into();
+        assert_eq!(tx.nonce(), &Felt::ZERO);
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v1().into();
+        assert_eq!(tx.nonce(), &Felt::from(7));
+
+        let tx: InvokeTransaction = dummy_tx_invoke_v3().into();
+        assert_eq!(tx.nonce(), &Felt::from(6));
+
+        let tx: DeclareTransaction = dummy_tx_declare_v0().into();
+        assert_eq!(tx.nonce(), &Felt::ZERO);
+
+        let tx: DeclareTransaction = dummy_tx_declare_v1().into();
+        assert_eq!(tx.nonce(), &Felt::from(5));
+
+        let tx: DeclareTransaction = dummy_tx_declare_v2().into();
+        assert_eq!(tx.nonce(), &Felt::from(6));
+
+        let tx: DeclareTransaction = dummy_tx_declare_v3().into();
+        assert_eq!(tx.nonce(), &Felt::from(5));
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v1().into();
+        assert_eq!(tx.nonce(), &Felt::from(4));
+
+        let tx: DeployAccountTransaction = dummy_tx_deploy_account_v3().into();
+        assert_eq!(tx.nonce(), &Felt::from(3));
+    }
+
+    #[test]
+    fn test_msg_to_l1_handler() {
+        let msg = starknet_core::types::MsgFromL1 {
+            from_address: starknet_core::types::EthAddress::from_hex("0x0000000000000000000000000000000000000001")
+                .unwrap(),
+            to_address: Felt::from(2),
+            entry_point_selector: Felt::from(3),
+            payload: vec![Felt::from(4), Felt::from(5)],
+        };
+
+        let l1_handler_expected = L1HandlerTransaction {
+            version: Felt::ZERO,
+            nonce: 0,
+            contract_address: Felt::from(2),
+            entry_point_selector: Felt::from(3),
+            calldata: vec![Felt::from(1), Felt::from(4), Felt::from(5)],
+        };
+
+        assert_eq!(L1HandlerTransaction::from(msg), l1_handler_expected);
+    }
+
+    #[test]
+    fn test_resource_bounds_mapping_conversion() {
+        let resource_mapping = ResourceBoundsMapping {
+            l1_gas: ResourceBounds { max_amount: 1, max_price_per_unit: 2 },
+            l2_gas: ResourceBounds { max_amount: 3, max_price_per_unit: 4 },
+        };
+
+        let starknet_resource_mapping: starknet_core::types::ResourceBoundsMapping = resource_mapping.clone().into();
+        let resource_mapping_back: ResourceBoundsMapping = starknet_resource_mapping.into();
+
+        assert_eq!(resource_mapping, resource_mapping_back);
+    }
+
+    #[test]
+    fn test_data_availability_mode_conversion() {
+        let da_mode = DataAvailabilityMode::L1;
+        let starknet_da_mode: starknet_core::types::DataAvailabilityMode = da_mode.into();
+        let da_mode_back: DataAvailabilityMode = starknet_da_mode.into();
+
+        assert_eq!(da_mode, da_mode_back);
+    }
+
+    pub(crate) fn dummy_tx_invoke_v0() -> InvokeTransactionV0 {
+        InvokeTransactionV0 {
+            max_fee: Felt::from(1),
+            signature: vec![Felt::from(2), Felt::from(3)],
+            contract_address: Felt::from(4),
+            entry_point_selector: Felt::from(5),
+            calldata: vec![Felt::from(6), Felt::from(7)],
+        }
+    }
+
+    pub(crate) fn dummy_tx_invoke_v1() -> InvokeTransactionV1 {
+        InvokeTransactionV1 {
+            sender_address: Felt::from(1),
+            calldata: vec![Felt::from(2), Felt::from(3)],
+            max_fee: Felt::from(4),
+            signature: vec![Felt::from(5), Felt::from(6)],
+            nonce: Felt::from(7),
+        }
+    }
+
+    pub(crate) fn dummy_tx_invoke_v3() -> InvokeTransactionV3 {
+        InvokeTransactionV3 {
+            sender_address: Felt::from(1),
+            calldata: vec![Felt::from(2), Felt::from(3)],
+            signature: vec![Felt::from(4), Felt::from(5)],
+            nonce: Felt::from(6),
+            resource_bounds: ResourceBoundsMapping {
+                l1_gas: ResourceBounds { max_amount: 1, max_price_per_unit: 2 },
+                l2_gas: ResourceBounds { max_amount: 3, max_price_per_unit: 4 },
+            },
+            tip: 7,
+            paymaster_data: vec![Felt::from(8), Felt::from(9)],
+            account_deployment_data: vec![Felt::from(10), Felt::from(11)],
+            nonce_data_availability_mode: DataAvailabilityMode::L1,
+            fee_data_availability_mode: DataAvailabilityMode::L2,
+        }
+    }
+
+    pub(crate) fn dummy_l1_handler() -> L1HandlerTransaction {
+        L1HandlerTransaction {
+            version: Felt::from(1),
+            nonce: 2,
+            contract_address: Felt::from(3),
+            entry_point_selector: Felt::from(4),
+            calldata: vec![Felt::from(5), Felt::from(6)],
+        }
+    }
+
+    pub(crate) fn dummy_tx_declare_v0() -> DeclareTransactionV0 {
+        DeclareTransactionV0 {
+            sender_address: Felt::from(1),
+            max_fee: Felt::from(2),
+            signature: vec![Felt::from(3), Felt::from(4)],
+            class_hash: Felt::from(5),
+        }
+    }
+
+    pub(crate) fn dummy_tx_declare_v1() -> DeclareTransactionV1 {
+        DeclareTransactionV1 {
+            sender_address: Felt::from(1),
+            max_fee: Felt::from(2),
+            signature: vec![Felt::from(3), Felt::from(4)],
+            nonce: Felt::from(5),
+            class_hash: Felt::from(6),
+        }
+    }
+
+    pub(crate) fn dummy_tx_declare_v2() -> DeclareTransactionV2 {
+        DeclareTransactionV2 {
+            sender_address: Felt::from(1),
+            compiled_class_hash: Felt::from(2),
+            max_fee: Felt::from(3),
+            signature: vec![Felt::from(4), Felt::from(5)],
+            nonce: Felt::from(6),
+            class_hash: Felt::from(7),
+        }
+    }
+
+    pub(crate) fn dummy_tx_declare_v3() -> DeclareTransactionV3 {
+        DeclareTransactionV3 {
+            sender_address: Felt::from(1),
+            compiled_class_hash: Felt::from(2),
+            signature: vec![Felt::from(3), Felt::from(4)],
+            nonce: Felt::from(5),
+            class_hash: Felt::from(6),
+            resource_bounds: ResourceBoundsMapping {
+                l1_gas: ResourceBounds { max_amount: 1, max_price_per_unit: 2 },
+                l2_gas: ResourceBounds { max_amount: 3, max_price_per_unit: 4 },
+            },
+            tip: 7,
+            paymaster_data: vec![Felt::from(8), Felt::from(9)],
+            account_deployment_data: vec![Felt::from(10), Felt::from(11)],
+            nonce_data_availability_mode: DataAvailabilityMode::L1,
+            fee_data_availability_mode: DataAvailabilityMode::L2,
+        }
+    }
+
+    pub(crate) fn dummy_tx_deploy() -> DeployTransaction {
+        DeployTransaction {
+            version: Felt::from(1),
+            contract_address_salt: Felt::from(2),
+            constructor_calldata: vec![Felt::from(3), Felt::from(4)],
+            class_hash: Felt::from(5),
+        }
+    }
+
+    pub(crate) fn dummy_tx_deploy_account_v1() -> DeployAccountTransactionV1 {
+        DeployAccountTransactionV1 {
+            max_fee: Felt::from(1),
+            signature: vec![Felt::from(2), Felt::from(3)],
+            nonce: Felt::from(4),
+            contract_address_salt: Felt::from(5),
+            constructor_calldata: vec![Felt::from(6), Felt::from(7)],
+            class_hash: Felt::from(8),
+        }
+    }
+
+    pub(crate) fn dummy_tx_deploy_account_v3() -> DeployAccountTransactionV3 {
+        DeployAccountTransactionV3 {
+            signature: vec![Felt::from(1), Felt::from(2)],
+            nonce: Felt::from(3),
+            contract_address_salt: Felt::from(4),
+            constructor_calldata: vec![Felt::from(5), Felt::from(6)],
+            class_hash: Felt::from(7),
+            resource_bounds: ResourceBoundsMapping {
+                l1_gas: ResourceBounds { max_amount: 1, max_price_per_unit: 2 },
+                l2_gas: ResourceBounds { max_amount: 3, max_price_per_unit: 4 },
+            },
+            tip: 8,
+            paymaster_data: vec![Felt::from(9), Felt::from(10)],
+            nonce_data_availability_mode: DataAvailabilityMode::L1,
+            fee_data_availability_mode: DataAvailabilityMode::L2,
         }
     }
 }